+++ Source/Buttons/continue.js
@@ -1,7 +1,7 @@
 const { getAdventure, endRoom } = require('../adventureDAO.js');
 const Button = require('../../Classes/Button.js');
 const { MessageActionRow, MessageButton } = require('discord.js');
-const { SAFE_DELIMITER } = require('../../helpers.js');
+const { SAFE_DELIMITER } = require('../../constants.js');
 
 const id = "continue";
 module.exports = new Button(id, (interaction, args) => {
