+++ Source/Buttons/join.js
@@ -1,6 +1,7 @@
 const Button = require('../../Classes/Button.js');
 const Delver = require('../../Classes/Delver.js');
-const { isSponsor, maxDelverCount } = require('../../helpers.js');
+const { maxDelverCount } = require("../../constants.js");
+const { isSponsor } = require('../../helpers.js');
 const { getGuild } = require('../guildDAO.js');
 const { getAdventure, setAdventure } = require("./../adventureDAO.js");
