+++ Source/Buttons/readyconsumable.js
@@ -2,7 +2,7 @@ const Button = require('../../Classes/Button.js');
 const { MessageEmbed, MessageActionRow, MessageSelectMenu } = require('discord.js');
 const { getColor } = require('../elementHelpers.js');
 const { getAdventure } = require('../adventureDAO.js');
-const { SAFE_DELIMITER } = require('../../helpers.js');
+const { SAFE_DELIMITER, MAX_SELECT_OPTIONS } = require('../../constants.js');
 const { getConsumable } = require('../consumables/_consumablesDictionary.js');
 
 const id = "readyconsumable";
@@ -23,7 +23,7 @@ module.exports = new Button(id, (interaction, args) => {
 					new MessageActionRow().addComponents(
 						new MessageSelectMenu().setCustomId(`consumable${SAFE_DELIMITER}${adventure.room.round}`)
 							.setPlaceholder("Pick a consumable...")
-							.addOptions(Object.keys(adventure.consumables).slice(0, 25).reduce((options, consumable) => options.concat({
+							.addOptions(Object.keys(adventure.consumables).slice(0, MAX_SELECT_OPTIONS).reduce((options, consumable) => options.concat({
 								label: `${consumable} (Held: ${adventure.consumables[consumable]})`,
 								description: getConsumable(consumable).description,
 								value: consumable
