+++ Source/Commands/invite.js
@@ -1,6 +1,6 @@
 const { MessageActionRow, MessageButton } = require('discord.js');
 const Command = require('../../Classes/Command.js');
-const { SAFE_DELIMITER } = require('../../helpers.js');
+const { SAFE_DELIMITER } = require('../../constants.js');
 const { getAdventure } = require('../adventureDAO.js');
 
 const id = "invite";
