+++ Source/Selects/loot.js
@@ -2,7 +2,7 @@ const { MessageActionRow, MessageButton } = require('discord.js');
 const Select = require('../../Classes/Select.js');
 const { getAdventure, updateRoomHeader, setAdventure } = require('../adventureDAO.js');
 const { getEquipmentProperty } = require('../equipment/_equipmentDictionary.js');
-const { SAFE_DELIMITER } = require('../../helpers.js');
+const { SAFE_DELIMITER } = require('../../constants.js');
 const { generateLootRow, generateRoutingRow } = require("../roomDAO.js");
 
 const id = "loot";
