+++ Source/equipment/firecracker-mercurial.js
@@ -1,6 +1,6 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier } = require('../combatantDAO.js');
-const { SAFE_DELIMITER } = require("../../helpers.js");
+const { SAFE_DELIMITER } = require("../../constants.js");
 
 module.exports = new EquipmentTemplate("Mercurial Firecracker", 2, "*Strike 3 random foes for @{damage} damage matching the user's element*\nCritical Hit: Damage x@{critBonus}", "Fire", effect, ["Double Firecracker", "Toxic Firecracker"])
 	.setCategory("Weapon")
