+++ bot.js
@@ -7,7 +7,7 @@ const { Routes } = require("discord-api-types/v9");
 const GuildProfile = require("./Classes/GuildProfile");
 const versionData = require('./Config/versionData.json');
 
-const { SAFE_DELIMITER } = require('./helpers.js');
+const { SAFE_DELIMITER } = require('./constants.js');
 const { loadAdventures } = require("./Source/adventureDAO.js");
 const { loadGuilds, setGuild } = require("./Source/guildDAO.js");
 const { loadPlayers } = require("./Source/playerDAO.js");
