+++ constants.js
@@ -1,0 +1,5 @@
+module.exports.SAFE_DELIMITER = "→";
+module.exports.MAX_MESSAGE_ACTION_ROWS = 5;
+module.exports.MAX_BUTTONS_PER_ROW = 5;
+module.exports.MAX_SELECT_OPTIONS = 25;
+module.exports.maxDelverCount = 8;
